refactor(array): move array type aliases
Move array type aliases to outer scope to make them available for other
files.
