+++ goesdr/array.py
